security

+++ src/main/java/com/example/hibernatecrud_spring/repository/EventRepository.java
@@ -2,5 +2,5 @@ package com.example.hibernatecrud_spring.repository;
 
 import com.example.hibernatecrud_spring.model.Event;
 
-public interface EventRepository extends GenericRepository<Event,Long> {
+public interface EventRepository extends GenericRepository<Event,  Integer> {
 }
