+++ src/main/java/com/example/hibernatecrud_spring/repository/FileRepository.java
@@ -2,5 +2,5 @@ package com.example.hibernatecrud_spring.repository;
 
 import com.example.hibernatecrud_spring.model.File;
 
-public interface FileRepository extends GenericRepository<File,Long> {
+public interface FileRepository extends GenericRepository<File,Integer> {
 }
